+++ sources/plugins/MegatecUps/changelog.md
@@ -1,3 +1,8 @@
+### 1.4.0
+* Fix disabling auto-test
+* Add detection of system shutdown caused by battery dead
+* Add daily auto-test time in configuration
+
 ### 1.3.0
 * Fix #562  : update usage of CDataContainer for better memory optimizations